Changed the names of the previousValue and finalValue variables in the MyIntEditTextOnFocusChangeListener class and all its implementations. The new names are oldValue and newValue.
Added two new properties to MyIntEditTextOnFocusChangeListener: oldString and newString. These work similarly to the oldValue and newValue but will always retain the actual string value of the text contents, rather than the parsed integers or any control values that are used as flags in the case of newValue and oldValue.

Added a method (printVariables) to print out a whole slew of info on the listener for a particular EditText. The four properties mentioned above along with its text contents at the time the method is called.

Streamlined the process of the onFocusChange() method so that parsing only happens when it needs to (when the EditText text being processed is not an empty string and when the newString is different from the oldString). In certain cases, the onFocusChange() method may finish early without having to execute through unnecessary code that involves parsing and validating input.

Empty string input is no longer handled by examining the error message from the exception thrown by trying to parse it.

Added tags to the EditText views in the NetworkConfigActivity via xml in activity_network_config.xml. Plan to add tags to the other EditText views before the next commit.

Changed the behavior of the valueChanged() method of the OnFocusChangeListener for the EditText views in the NetworkConfigActivity. Not inputting anything now causes the field to revert back to its previous contents, rather than clearing the view.