Add new EditText subclasses
A set of new subclasses to the [EditText] view were added to the project
as a means to implement the code currently contained within the
[MyIntEditTextOnFocusChangeListener]. The first [InputField] is abstract
and sets the framework for implementing the [OnFocusChangeListener].
It has a non-abstract subclass, [IntegerField] which is meant to properly
implement all the code currently found in the custom listener class.
This in turn has one subclass [AddressField] which currently is identical
to its parent.