Implement [RuleSet] static inner class
Finally found a good solution for the situation of custom handling of the
different events that may be triggered by attempting to change the text
of an [IntegerField]. A set of simple rules in the class itself still streamlines
the process of determining what kind of change was made and whether it
warrants additional validation. A new instance variable in the class is a
reference to a [RuleSet] object which contains methods similar to those
in the [View] event listener classes we have been using. Multiple fields
can have the same [RuleSet] instance assigned to them and have things
within handled as necessary. This is curerntly implemented in the
[NetworkConfigActivity] class. Next step will be to clean up some no-
longer-needed functionality and variables in this process.